requirements.txt
Babel==2.3.4
dj-database-url==0.4.1
Django==1.9.5
django-bootstrap3==7.0.1
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-phonenumber-field==1.1.0
django-rest-swagger==0.3.5
django-sendgrid==1.0.1
django-twilio==0.8.0
djangorestframework==3.3.3
gunicorn==19.4.5
httplib2==0.9.2
oauth2==1.9.0.post1
phonenumberslite==7.3.2
Pillow==3.2.0
postgres==2.2.1
psycopg2==2.6.1
PySocks==1.5.6
python-dateutil==2.5.2
python-http-client==1.2.3
pytz==2016.4
PyYAML==3.11
requests==2.9.1
sendgrid==2.2.1
sendgrid-django==2.0.0
six==1.10.0
smtpapi==0.3.1
sqlparse==0.1.19
stripe==1.32.2
twilio==5.4.0
Werkzeug==0.11.5
whitenoise==3.0